Fix compilation warning on 32-bit.
Former-commit-id: 1ec11054c849bdc0fbe3b85d6696b510638c33b1

+++ server/src/server/utils.c
@@ -63,7 +63,7 @@ int rndm_chance(uint32 n)
 		return 0;
 	}
 
-	return RANDOM() < (RAND_MAX + 1U) / n;
+	return (uint32) RANDOM() < (RAND_MAX + 1U) / n;
 }
 
 /**
